+++ mozrepl/type/function.py
@@ -4,7 +4,7 @@
 from __future__ import unicode_literals, absolute_import, division, print_function
 import itertools
 
-from . import Object
+from .object import Object
 
 class Function(Object):
 	"""
@@ -18,18 +18,14 @@ class Function(Object):
 	
 	def __call__(self, *args, **kwargs):
 		"""
-		오직 문자열과 숫자만 전달 가능.
+		내부에서 :py:function:`~mozrepl.util.convertToJs` 함수를 호출하여 처리합니다. 입력하는 각 인자는 이 함수에서 허용하는 형식을 준수해야 합니다.
 		
-		:todo: 함수 및 오브젝트를 전달 가능하도록 만들기. 문자열에 \'문자가 포함되어 있을 경우 문제가 발생 할 수 있음. 콰우팅 시켜주기.
-			None 타입의 경우 null로 처리하도록 함.
+		:todo: 
 			+ mozrepl.type.Function에서 발생하던 'TypeError: context is undefined' 오류를 수정. [`tb69wn6127`_]
 		"""
 		v = list()
 		for i in itertools.chain(args, kwargs.values()):
-			if isinstance(i, int):
-				v.append(str(i))
-			elif isinstance(i, unicode):
-				v.append(repr(str(i)))
+			type
 		buffer = '{0}({1})'.format(self.reference, ','.join(v))
 		print(buffer)
 		return self._repl.execute(buffer)